+++ react/index.tsx
@@ -138,7 +138,7 @@ function ShippingOptionZipcode({
         selectedZipcode={selectedZipcode}
         onSubmit={(zipCode: string) => {
           setWasLocationModalOpenedByEvent(true)
-          onSubmit(zipCode, false)
+          onSubmit(zipCode, true)
         }}
         inputErrorMessage={submitErrorMessage?.message}
         callToAction={callToAction}
@@ -171,7 +171,7 @@ function ShippingOptionZipcode({
         onClose={() => setIsLocationModalOpen(false)}
         showLocationDetectorButton={showLocationDetectorButton}
         onSubmit={async (zipcode: string) => {
-          onSubmit(zipcode, false)
+          onSubmit(zipcode, true)
         }}
         isLoading={isLoading}
         inputErrorMessage={submitErrorMessage}
@@ -192,7 +192,7 @@ function ShippingOptionZipcode({
         onPickupMethodSelection={onPickupMethodSelection}
         pickupProps={{
           onSelectPickup,
-          onSubmit: (value) => onSubmit(value, false),
+          onSubmit: (value) => onSubmit(value, true),
           pickups,
           inputErrorMessage: submitErrorMessage?.message,
           selectedPickup: pickup,
@@ -207,7 +207,7 @@ function ShippingOptionZipcode({
         onClose={() => setIsPickupModalOpen(false)}
         pickupProps={{
           onSelectPickup,
-          onSubmit: (value) => onSubmit(value, false),
+          onSubmit: (value) => onSubmit(value, true),
           pickups,
           inputErrorMessage: submitErrorMessage?.message,
           selectedPickup: pickup,