+++ CHANGELOG.md
@@ -14,6 +14,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Shipping method modal no longer opens automatically and can be dismissed without selection
 - Removed auto select pickup
 
+## [1.11.0] - 2025-12-09
+
+### Changed
+
+- Not forwarding cookies to `getCatalogCount` anymore
+- Refresh page after `zipcode update` at `segment cookie`
+
 ## [1.10.0] - 2025-10-28
 
 ### Added